requirements.txt
# Flask web framework
flask==2.3.3
flask-wtf==1.2.1

# Database
sqlalchemy==2.0.23

# Microsoft Graph API
requests==2.31.0
msal==1.24.1
geopy==2.4.1

# Email sending
# Using standard library's smtplib instead of sendgrid

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
